Lockhammer: Assign thread affinity prior to creation
Set thread affinity to a particular hardware thread prior to thread
creation.  This creates a potential disconnect between the start
order and the cpu on which the thread is running necessitating some
changes to the synchronized start code and the addition of a core
number thread argument in order to correctly index per-cpu variables
inside the locks algorithms.

Fixes #16

Change-Id: Ia27d70e7d2875637a4ef1514e61636fc1b662698